Maven: Move reusable part to new class MavenSupport
Move the management of the Maven container and session as well as factory
functions to the new class MavenSupport. This is a preparation for using
Maven to get the metadata for Gradle dependencies.
