+++ analyzer/src/main/kotlin/managers/Maven.kt
@@ -21,9 +21,10 @@ package com.here.ort.analyzer.managers
 
 import ch.frankel.slf4k.*
 
-import com.here.ort.analyzer.MavenLogger
+import com.here.ort.analyzer.MavenSupport
 import com.here.ort.analyzer.PackageManager
 import com.here.ort.analyzer.PackageManagerFactory
+import com.here.ort.analyzer.identifier
 import com.here.ort.downloader.VersionControlSystem
 import com.here.ort.model.AnalyzerResult
 import com.here.ort.model.Package
@@ -34,25 +35,11 @@ import com.here.ort.util.log
 
 import java.io.File
 
-import org.apache.maven.artifact.repository.LegacyLocalRepositoryManager
 import org.apache.maven.bridge.MavenRepositorySystem
-import org.apache.maven.execution.DefaultMavenExecutionRequest
 import org.apache.maven.model.building.ModelBuildingRequest
-import org.apache.maven.project.DefaultProjectBuildingRequest
-import org.apache.maven.project.MavenProject
 import org.apache.maven.project.ProjectBuilder
 import org.apache.maven.project.ProjectBuildingResult
-import org.apache.maven.repository.internal.MavenRepositorySystemUtils
 
-import org.codehaus.plexus.DefaultContainerConfiguration
-import org.codehaus.plexus.DefaultPlexusContainer
-import org.codehaus.plexus.PlexusConstants
-import org.codehaus.plexus.PlexusContainer
-import org.codehaus.plexus.classworlds.ClassWorld
-import org.codehaus.plexus.logging.BaseLoggerManager
-
-import org.eclipse.aether.DefaultRepositorySystemSession
-import org.eclipse.aether.RepositorySystem
 import org.eclipse.aether.RepositorySystemSession
 import org.eclipse.aether.artifact.Artifact
 import org.eclipse.aether.graph.DependencyNode
@@ -73,8 +60,6 @@ class Maven : PackageManager() {
             "Java",
             listOf("pom.xml")
     ) {
-        val SCM_REGEX = Regex("scm:[^:]+:(.+)")
-
         override fun create() = Maven()
     }
 
@@ -84,15 +69,17 @@ class Maven : PackageManager() {
      */
     private val assumedNonDeliveredScopes = setOf("test")
 
-    private val container = createContainer()
-    private val repositorySystemSession = createRepositorySystemSession()
+    private val maven = MavenSupport { localRepositoryManager ->
+        LocalRepositoryManagerWrapper(localRepositoryManager)
+    }
+
     private val projectsByIdentifier = mutableMapOf<String, ProjectBuildingResult>()
 
     override fun command(workingDir: File) = "mvn"
 
     override fun prepareResolution(definitionFiles: List<File>) {
-        val projectBuilder = container.lookup(ProjectBuilder::class.java, "default")
-        val projectBuildingRequest = createProjectBuildingRequest(false)
+        val projectBuilder = maven.container.lookup(ProjectBuilder::class.java, "default")
+        val projectBuildingRequest = maven.createProjectBuildingRequest(false)
         val projectBuildingResults = projectBuilder.build(definitionFiles, false, projectBuildingRequest)
 
         projectBuildingResults.forEach { projectBuildingResult ->
@@ -104,7 +91,7 @@ class Maven : PackageManager() {
     }
 
     override fun resolveDependencies(projectDir: File, workingDir: File, definitionFile: File): AnalyzerResult? {
-        val projectBuildingResult = buildMavenProject(definitionFile)
+        val projectBuildingResult = maven.buildMavenProject(definitionFile)
         val mavenProject = projectBuildingResult.project
         val packages = mutableMapOf<String, Package>()
         val scopes = mutableMapOf<String, Scope>()
@@ -118,10 +105,10 @@ class Maven : PackageManager() {
             scope.dependencies.add(parseDependency(node, packages))
         }
 
-        val vcsInfo = parseVcsInfo(mavenProject).let {
+        val vcsInfo = maven.parseVcsInfo(mavenProject).let {
             if (it.isEmpty()) {
                 val vcs = VersionControlSystem.fromDirectory(projectDir)
-                VcsInfo(vcs?.toString() ?: "", vcs?.getRemoteUrl(projectDir) ?: "",
+                MavenSupport.VcsInfo(vcs?.toString() ?: "", vcs?.getRemoteUrl(projectDir) ?: "",
                         vcs?.getWorkingRevision(projectDir) ?: "")
             } else it
         }
@@ -143,51 +130,6 @@ class Maven : PackageManager() {
         return AnalyzerResult(true, project, packages.values.toSortedSet())
     }
 
-    private fun buildMavenProject(pomFile: File): ProjectBuildingResult {
-        val projectBuilder = container.lookup(ProjectBuilder::class.java, "default")
-        val projectBuildingRequest = createProjectBuildingRequest(true)
-
-        return projectBuilder.build(pomFile, projectBuildingRequest)
-    }
-
-    private fun createContainer(): PlexusContainer {
-        val configuration = DefaultContainerConfiguration().apply {
-            autoWiring = true
-            classPathScanning = PlexusConstants.SCANNING_INDEX
-            classWorld = ClassWorld("plexus.core", javaClass.classLoader)
-        }
-
-        return DefaultPlexusContainer(configuration).apply {
-            loggerManager = object : BaseLoggerManager() {
-                override fun createLogger(name: String) = MavenLogger(log.effectiveLevel)
-            }
-        }
-    }
-
-    private fun createProjectBuildingRequest(resolveDependencies: Boolean) =
-            DefaultProjectBuildingRequest().apply {
-                isResolveDependencies = resolveDependencies
-                repositorySession = repositorySystemSession
-                systemProperties["java.home"] = System.getProperty("java.home")
-                systemProperties["java.version"] = System.getProperty("java.version")
-            }
-
-
-    private fun createRepositorySystemSession(): RepositorySystemSession {
-        val mavenRepositorySystem = container.lookup(MavenRepositorySystem::class.java, "default")
-        val aetherRepositorySystem = container.lookup(RepositorySystem::class.java, "default")
-        val repositorySystemSession = MavenRepositorySystemUtils.newSession()
-        val mavenExecutionRequest = DefaultMavenExecutionRequest()
-        val localRepository = mavenRepositorySystem.createLocalRepository(mavenExecutionRequest,
-                org.apache.maven.repository.RepositorySystem.defaultUserLocalRepository)
-
-        val session = LegacyLocalRepositoryManager.overlay(localRepository, repositorySystemSession,
-                aetherRepositorySystem)
-        val wrapper = LocalRepositoryManagerWrapper(session.localRepositoryManager)
-
-        return DefaultRepositorySystemSession(session).setLocalRepositoryManager(wrapper)
-    }
-
     private fun parseDependency(node: DependencyNode, packages: MutableMap<String, Package>): PackageReference {
         val pkg = packages.getOrPut(node.artifact.identifier()) {
             parsePackage(node)
@@ -199,9 +141,9 @@ class Maven : PackageManager() {
     }
 
     private fun parsePackage(node: DependencyNode): Package {
-        val mavenRepositorySystem = container.lookup(MavenRepositorySystem::class.java, "default")
-        val projectBuilder = container.lookup(ProjectBuilder::class.java, "default")
-        val projectBuildingRequest = createProjectBuildingRequest(true).apply {
+        val mavenRepositorySystem = maven.container.lookup(MavenRepositorySystem::class.java, "default")
+        val projectBuilder = maven.container.lookup(ProjectBuilder::class.java, "default")
+        val projectBuildingRequest = maven.createProjectBuildingRequest(true).apply {
             validationLevel = ModelBuildingRequest.VALIDATION_LEVEL_MINIMAL
         }
 
@@ -226,31 +168,12 @@ class Maven : PackageManager() {
                 downloadUrl = "", // TODO: Try to get URL for downloaded dependencies.
                 hash = "", // TODO: Get hash from local metadata?
                 hashAlgorithm = "",
-                vcsProvider = parseVcsProvider(mavenProject),
-                vcsUrl = parseVcsUrl(mavenProject),
-                vcsRevision = parseVcsRevision(mavenProject),
+                vcsProvider = maven.parseVcsProvider(mavenProject),
+                vcsUrl = maven.parseVcsUrl(mavenProject),
+                vcsRevision = maven.parseVcsRevision(mavenProject),
                 vcsPath = ""
         )
     }
-
-    private fun parseVcsInfo(mavenProject: MavenProject) =
-            VcsInfo(parseVcsProvider(mavenProject), parseVcsUrl(mavenProject), parseVcsRevision(mavenProject))
-
-    private fun parseVcsProvider(mavenProject: MavenProject): String {
-        mavenProject.scm?.connection?.split(":")?.let {
-            if (it.size > 1 && it[0] == "scm") {
-                return it[1]
-            }
-        }
-        return ""
-    }
-
-    private fun parseVcsUrl(mavenProject: MavenProject) =
-            mavenProject.scm?.connection?.let {
-                SCM_REGEX.matchEntire(it)?.groupValues?.getOrNull(1)
-            } ?: ""
-
-    private fun parseVcsRevision(mavenProject: MavenProject) = mavenProject.scm?.tag ?: ""
 
     /**
      * A wrapper for the [LocalRepositoryManager] used in [repositorySystemSession] that pretends that the POM files of
@@ -300,9 +223,4 @@ class Maven : PackageManager() {
         override fun getRepository(): LocalRepository = localRepositoryManager.repository
     }
 
-    private data class VcsInfo(val provider: String, val url: String, val revision: String) {
-        fun isEmpty() = provider.isEmpty() && url.isEmpty() && revision.isEmpty()
-    }
-
-    fun Artifact.identifier() = "$groupId:$artifactId:$version"
 }
